refactor(calendario): split CalendarioObras + CalendarioCompras (Bloco 4.4 — pages)
Split incremental das duas páginas-monstro mantendo todos os comportamentos
existentes (URL persistence, 5 view modes em obras, filtros bespoke):

CalendarioObras (1011 → 348 linhas no orquestrador, -66%):
- src/pages/CalendarioObras/types.ts: ViewMode, status badges, periodLabel,
  getActivityStatus.
- useCalendarObrasUrlState.ts: hook que centraliza view + refDate + range +
  filtros + navegação prev/next/today + draft range — toda a persistência
  na query string num único lugar.
- CalendarObrasToolbar.tsx: tabs de view, badges de contagem, navegador de
  período (com bloco especial pra range com aplicar/resetar).
- CalendarObrasFilters.tsx: filtros de obra/etapa/concluídas isolados.
- WeekListView.tsx: visão "Semana · Lista" extraída do antigo helper inline.
- index.tsx orquestra os pedaços + componentes especialistas
  (CalendarMonthGrid, CalendarDayAgenda, CalendarRangeTimeline) e os
  dialogs (ActivityDetailDialog, BreakActivityDialog, NonWorkingDaysDialog).

CalendarioCompras (1195 → 721 linhas, -40%):
- src/pages/calendario-compras/types.ts: PurchaseWithProject, CalendarStatus,
  calendarStatusConfig, formatters (fmt, fmtCompact, fmtDiff,
  fmtRequestedDate), toCalendarStatus, ProjectPurchaseInsert.
- cells.tsx: DateCell, StatusCell, ActualCostCell e PurchaseRowDetail —
  todas as células editáveis e o bloco de detalhe expandido em um arquivo
  coeso.
- NewPurchaseDialog.tsx: modal de criação de solicitação completo,
  incluindo validações campo-a-campo e fallback de prazo.

Página principal de Compras importa esses primitivos e mantém a lógica
de listagem/calendário, filtros, mutations e ordenação. Não realocada
para a folder pattern Cronograma/PainelObras porque o monolito ainda
contém ~700 linhas de view (calendar grid + 2 tabelas com agrupamentos),
o que é tema do próximo PR.

Notas honestas:
- Meta literal "< 250 linhas cada página" da issue não atingida: o conjunto
  de view modes (5 em obras), filtros bespoke e mutations específicas pesa
  além do que cabe. Os 348/721 LOC entregues representam orquestração
  enxuta — cada sub-componente é independente e testável.
- ProjectCalendar (entregue em c65cd88) não é adotado nestas páginas
  porque CalendarMonthGrid já é deeply activity-aware (etapa anterior
  atrasada, replanejamento, business days). Migração específica é
  follow-up dedicado.

Refs #21
