+++ src/pages/CalendarioCompras.tsx
@@ -1,14 +1,13 @@
-import { Fragment, useMemo, useState, useEffect, useRef } from 'react';
+import { Fragment, useMemo, useState } from 'react';
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
 import {
   format, parseISO, startOfMonth, endOfMonth, eachDayOfInterval,
-  isSameDay, isSameMonth, addMonths, subMonths, isWeekend, addDays, isValid,
+  isSameDay, isSameMonth, addMonths, subMonths, isWeekend,
 } from 'date-fns';
 import { ptBR } from 'date-fns/locale';
 import {
-  ChevronLeft, ChevronRight, Calendar, Check, X, Pencil, CalendarIcon,
-  FilterX, Plus, ChevronDown, ChevronUp, ExternalLink, Package,
-  FileText, Truck, ArrowUpDown,
+  ChevronLeft, ChevronRight, Calendar, CalendarIcon, X, FilterX, Plus, ChevronDown, ChevronUp,
+  ExternalLink, ArrowUpDown,
 } from 'lucide-react';
 import { useNavigate } from 'react-router-dom';
 import { toast } from 'sonner';
@@ -17,6 +16,9 @@ import { Button } from '@/components/ui/button';
 import { Badge } from '@/components/ui/badge';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import { Input } from '@/components/ui/input';
+import { Label } from '@/components/ui/label';
+import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover';
+import { Calendar as CalendarPicker } from '@/components/ui/calendar';
 import {
   Select, SelectContent, SelectItem, SelectTrigger, SelectValue,
 } from '@/components/ui/select';
@@ -24,503 +26,27 @@ import {
   Table, TableBody, TableCell, TableHead, TableHeader, TableRow,
 } from '@/components/ui/table';
 import { PageSkeleton } from '@/components/ui-premium';
-import { Calendar as CalendarPicker } from '@/components/ui/calendar';
-import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover';
-import {
-  Dialog, DialogContent, DialogHeader, DialogTitle, DialogFooter,
-} from '@/components/ui/dialog';
-import { Label } from '@/components/ui/label';
-import { Textarea } from '@/components/ui/textarea';
 import { cn } from '@/lib/utils';
 import { PageContainer } from '@/components/layout/PageContainer';
 import { PageHeader } from '@/components/layout/PageHeader';
-import type { ProjectPurchase } from '@/hooks/useProjectPurchases';
-import type { TablesInsert } from '@/integrations/supabase/types';
-import { useAuth } from '@/hooks/useAuth';
-import { Clock, ThumbsUp, CheckCircle2, AlertTriangle } from 'lucide-react';
+import {
+  type CalendarStatus,
+  type PurchaseWithProject,
+  CALENDAR_STATUS_OPTIONS,
+  calendarStatusConfig,
+  fmt,
+  fmtCompact,
+  fmtDiff,
+  fmtRequestedDate,
+  toCalendarStatus,
+} from './calendario-compras/types';
+import { ActualCostCell, DateCell, PurchaseRowDetail, StatusCell } from './calendario-compras/cells';
+import { NewPurchaseDialog } from './calendario-compras/NewPurchaseDialog';
 
-/**
- * Payload tipado para INSERT em `project_purchases`.
- * Deriva diretamente do schema gerado pelo Supabase, garantindo que qualquer
- * mudança de coluna (NOT NULL, novo campo, etc.) seja refletida em tempo de build.
- *
- * Uso:
- *   const payload: ProjectPurchaseInsert = { project_id, item_name, created_by, required_by_date, ... };
- *   await supabase.from('project_purchases').insert(payload);
- */
-export type ProjectPurchaseInsert = TablesInsert<'project_purchases'>;
+// `ProjectPurchaseInsert` re-exportado para compatibilidade com consumidores
+// externos que possam estar importando dessa página.
+export type { ProjectPurchaseInsert } from './calendario-compras/types';
 
-interface PurchaseWithProject extends Omit<ProjectPurchase, 'created_at'> {
-  project_name: string;
-  payment_due_date?: string | null;
-  /**
-   * Override defensivo: embora a coluna `created_at` seja NOT NULL no banco,
-   * registros antigos sincronizados ou casos de borda podem chegar sem o campo.
-   * Tratamos como opcional/nulo para forçar o uso do helper `fmtRequestedDate`,
-   * que já retorna "—" para null/undefined/strings inválidas.
-   */
-  created_at?: string | null;
-}
-
-type CalendarStatus = 'pending' | 'approved' | 'delivered' | 'delayed';
-
-/**
- * Tons das badges de status — usam variantes claras + escuras com `dark:` para
- * manter contraste AA tanto em light quanto em dark mode. Usar tokens semânticos
- * resolveria de forma mais limpa, mas mantemos a paleta cromática (amber/blue/
- * emerald/red) para preservar o reconhecimento visual já consolidado em outras
- * telas (Compras, Painel de Obras).
- */
-const calendarStatusConfig: Record<CalendarStatus, { label: string; color: string; icon: React.ElementType }> = {
-  pending:   { label: 'Pendente',  color: 'bg-amber-100 text-amber-800 border-amber-300 dark:bg-amber-500/15 dark:text-amber-300 dark:border-amber-500/40',           icon: Clock },
-  approved:  { label: 'Aprovado',  color: 'bg-blue-100 text-blue-800 border-blue-300 dark:bg-blue-500/15 dark:text-blue-300 dark:border-blue-500/40',                 icon: ThumbsUp },
-  delivered: { label: 'Entregue',  color: 'bg-emerald-100 text-emerald-800 border-emerald-300 dark:bg-emerald-500/15 dark:text-emerald-300 dark:border-emerald-500/40', icon: CheckCircle2 },
-  delayed:   { label: 'Atrasado',  color: 'bg-red-100 text-red-800 border-red-300 dark:bg-red-500/15 dark:text-red-300 dark:border-red-500/40',                       icon: AlertTriangle },
-};
-
-const CALENDAR_STATUS_OPTIONS: CalendarStatus[] = ['pending', 'approved', 'delivered', 'delayed'];
-
-function toCalendarStatus(s: string | null | undefined): CalendarStatus {
-  if (s === 'approved' || s === 'awaiting_approval' || s === 'purchased' || s === 'ordered' || s === 'in_transit') return 'approved';
-  if (s === 'delivered' || s === 'sent_to_site') return 'delivered';
-  if (s === 'delayed') return 'delayed';
-  return 'pending';
-}
-
-// ─── DateCell ────────────────────────────────────────────────────────────────
-function DateCell({
-  value, onSave, placeholder = 'Definir',
-}: { value: string | null | undefined; onSave: (v: string | null) => void; placeholder?: string }) {
-  const [open, setOpen] = useState(false);
-  const selected = value ? parseISO(value) : undefined;
-  return (
-    <Popover open={open} onOpenChange={setOpen}>
-      <PopoverTrigger asChild>
-        <button
-          type="button"
-          className={cn(
-            'group inline-flex items-center gap-1 rounded px-1 py-0.5 text-xs hover:bg-muted whitespace-nowrap',
-            !value && 'text-muted-foreground italic',
-          )}
-        >
-          {value ? format(parseISO(value), 'dd/MM/yy') : placeholder}
-          {value ? (
-            <X className="h-3 w-3 opacity-0 group-hover:opacity-60 hover:text-destructive transition-opacity"
-              onClick={(e) => { e.stopPropagation(); onSave(null); setOpen(false); }} />
-          ) : (
-            <Pencil className="h-3 w-3 opacity-0 group-hover:opacity-60 transition-opacity" />
-          )}
-        </button>
-      </PopoverTrigger>
-      <PopoverContent className="w-auto p-0 z-[200]" align="start">
-        <CalendarPicker
-          mode="single"
-          selected={selected}
-          onSelect={(d) => { if (d) { onSave(format(d, 'yyyy-MM-dd')); setOpen(false); } }}
-          locale={ptBR}
-          initialFocus
-          className="p-3 pointer-events-auto"
-        />
-      </PopoverContent>
-    </Popover>
-  );
-}
-
-// ─── StatusCell ───────────────────────────────────────────────────────────────
-function StatusCell({ purchase, onSave }: { purchase: PurchaseWithProject; onSave: (id: string, v: CalendarStatus) => void }) {
-  const current = toCalendarStatus(purchase.status);
-  const cfg = calendarStatusConfig[current];
-  const Icon = cfg.icon;
-  return (
-    <Select value={current} onValueChange={(v) => onSave(purchase.id, v as CalendarStatus)}>
-      <SelectTrigger
-        className={cn(
-          // h-7 + padding fino + cor do status. `[&>span]:line-clamp-none` evita
-          // que o SelectValue corte o label em uma linha muito apertada.
-          'h-7 min-w-[120px] w-auto text-[11px] pl-2 pr-1.5 py-0 gap-1 border rounded-full font-medium',
-          '[&>span]:line-clamp-none [&>span]:flex [&>span]:items-center [&>span]:gap-1',
-          '[&_svg]:opacity-70',
-          cfg.color,
-        )}
-        aria-label={`Status: ${cfg.label}`}
-      >
-        <span className="inline-flex items-center gap-1 whitespace-nowrap">
-          <Icon className="h-3 w-3 shrink-0" aria-hidden />
-          <SelectValue />
-        </span>
-      </SelectTrigger>
-      <SelectContent>
-        {CALENDAR_STATUS_OPTIONS.map((s) => {
-          const c = calendarStatusConfig[s];
-          const I = c.icon;
-          return (
-            <SelectItem key={s} value={s} className="text-xs">
-              <span className="inline-flex items-center gap-1.5"><I className="h-3 w-3" />{c.label}</span>
-            </SelectItem>
-          );
-        })}
-      </SelectContent>
-    </Select>
-  );
-}
-
-// ─── ActualCostCell ───────────────────────────────────────────────────────────
-const fmtCompact = (v: number | null) =>
-  v != null ? v.toLocaleString('pt-BR', { style: 'currency', currency: 'BRL', maximumFractionDigits: 0 }) : '—';
-const fmt = (v: number | null) =>
-  v != null ? v.toLocaleString('pt-BR', { style: 'currency', currency: 'BRL' }) : '—';
-const fmtDiff = (v: number) => {
-  const sign = v > 0 ? '+' : v < 0 ? '-' : '';
-  const abs = Math.abs(v).toLocaleString('pt-BR', { style: 'currency', currency: 'BRL', maximumFractionDigits: 0 });
-  return `${sign}${abs}`;
-};
-
-/**
- * Formata `created_at` (ISO UTC vindo do Supabase, ex: "2025-04-26T03:10:00+00:00")
- * sempre como `dd/MM/yyyy` no fuso horário LOCAL do navegador.
- *
- * - `parseISO` interpreta corretamente o offset `Z`/`+00:00`, evitando o bug clássico
- *   de `new Date("2025-04-26")` que assumiria UTC e poderia exibir 25/04 em -03:00.
- * - Retorna "—" para null/undefined/string inválida (defensivo contra dados antigos).
- */
-const fmtRequestedDate = (iso: string | null | undefined): string => {
-  if (!iso) return '—';
-  try {
-    const d = parseISO(iso);
-    if (!isValid(d)) return '—';
-    return format(d, 'dd/MM/yyyy', { locale: ptBR });
-  } catch {
-    return '—';
-  }
-};
-
-function ActualCostCell({ purchase, onSave }: { purchase: PurchaseWithProject; onSave: (id: string, v: number | null) => void }) {
-  const [editing, setEditing] = useState(false);
-  const [value, setValue] = useState(purchase.actual_cost != null ? String(purchase.actual_cost) : '');
-  useEffect(() => { setValue(purchase.actual_cost != null ? String(purchase.actual_cost) : ''); }, [purchase.actual_cost]);
-  const commit = () => {
-    const trimmed = value.trim().replace(',', '.');
-    const num = trimmed === '' ? null : Number(trimmed);
-    if (trimmed !== '' && (isNaN(num as number) || (num as number) < 0)) {
-      toast.error('Valor inválido'); setValue(purchase.actual_cost != null ? String(purchase.actual_cost) : ''); setEditing(false); return;
-    }
-    if (num !== purchase.actual_cost) onSave(purchase.id, num);
-    setEditing(false);
-  };
-  if (editing) {
-    return (
-      <div className="flex items-center gap-1">
-        <Input autoFocus type="number" step="0.01" min="0" inputMode="decimal" value={value}
-          onChange={(e) => setValue(e.target.value)}
-          onKeyDown={(e) => { if (e.key === 'Enter') commit(); if (e.key === 'Escape') { setValue(purchase.actual_cost != null ? String(purchase.actual_cost) : ''); setEditing(false); } }}
-          onBlur={commit} className="h-7 w-24 text-xs px-2" />
-        <Button size="icon" variant="ghost" className="h-6 w-6" onMouseDown={(e) => e.preventDefault()} onClick={commit}><Check className="h-3 w-3" /></Button>
-        <Button size="icon" variant="ghost" className="h-6 w-6" onMouseDown={(e) => e.preventDefault()} onClick={() => { setValue(purchase.actual_cost != null ? String(purchase.actual_cost) : ''); setEditing(false); }}><X className="h-3 w-3" /></Button>
-      </div>
-    );
-  }
-  return (
-    <button type="button" onClick={() => setEditing(true)}
-      className="group inline-flex items-center gap-1 rounded px-1 py-0.5 text-xs hover:bg-muted whitespace-nowrap">
-      <span className={cn(purchase.actual_cost == null && 'text-muted-foreground italic')}>
-        {purchase.actual_cost != null ? fmtCompact(purchase.actual_cost) : 'Informar'}
-      </span>
-      <Pencil className="h-3 w-3 opacity-0 group-hover:opacity-60 transition-opacity" />
-    </button>
-  );
-}
-
-// ─── Expandable row details ───────────────────────────────────────────────────
-function PurchaseRowDetail({ p }: { p: PurchaseWithProject }) {
-  return (
-    // O detalhe ocupa naturalmente a largura da célula colSpan. Em telas largas
-    // a tabela já é responsiva (sem scroll-x); em telas estreitas, o scroll
-    // horizontal da própria tabela leva o detail row junto, mantendo alinhamento.
-    // `max-w-screen-2xl` evita expansão visual exagerada quando há poucas colunas.
-    <div className="w-full">
-      <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-x-6 gap-y-3 px-4 py-3 text-xs bg-muted/30 border-t max-w-screen-2xl">
-        {p.category && (
-          <div className="min-w-0">
-            <span className="text-muted-foreground flex items-center gap-1 mb-0.5">
-              <Package className="h-3 w-3 shrink-0" /> Categoria
-            </span>
-            <p className="font-medium break-words">{p.category}</p>
-          </div>
-        )}
-        {p.supplier_name && (
-          <div className="min-w-0">
-            <span className="text-muted-foreground flex items-center gap-1 mb-0.5">
-              <Truck className="h-3 w-3 shrink-0" /> Fornecedor
-            </span>
-            <p className="font-medium break-words">{p.supplier_name}</p>
-          </div>
-        )}
-        {p.quantity && (
-          <div className="min-w-0">
-            <span className="text-muted-foreground flex items-center gap-1 mb-0.5">
-              <Package className="h-3 w-3 shrink-0" /> Qtde / Unidade
-            </span>
-            <p className="font-medium">{p.quantity}{p.unit ? ` ${p.unit}` : ''}</p>
-          </div>
-        )}
-        {p.delivery_address && (
-          <div className="min-w-0">
-            <span className="text-muted-foreground flex items-center gap-1 mb-0.5">
-              <Truck className="h-3 w-3 shrink-0" /> Entrega
-            </span>
-            <p className="font-medium break-words">{p.delivery_address}</p>
-          </div>
-        )}
-        {p.description && (
-          <div className="col-span-full min-w-0">
-            <span className="text-muted-foreground flex items-center gap-1 mb-0.5">
-              <FileText className="h-3 w-3 shrink-0" /> Descrição
-            </span>
-            <p className="text-foreground leading-snug break-words">{p.description}</p>
-          </div>
-        )}
-        {p.notes && (
-          <div className="col-span-full min-w-0">
-            <span className="text-muted-foreground mb-0.5 block">Observações</span>
-            <p className="text-foreground leading-snug break-words">{p.notes}</p>
-          </div>
-        )}
-      </div>
-    </div>
-  );
-}
-
-// ─── New Purchase Dialog ──────────────────────────────────────────────────────
-interface NewPurchaseForm {
-  project_id: string;
-  item_name: string;
-  category: string;
-  supplier_name: string;
-  estimated_cost: string;
-  planned_purchase_date: Date | undefined;
-  quantity: string;
-  unit: string;
-  description: string;
-  notes: string;
-}
-
-const EMPTY_FORM: NewPurchaseForm = {
-  project_id: '',
-  item_name: '',
-  category: '',
-  supplier_name: '',
-  estimated_cost: '',
-  planned_purchase_date: undefined,
-  quantity: '',
-  unit: '',
-  description: '',
-  notes: '',
-};
-
-function NewPurchaseDialog({
-  open,
-  onClose,
-  projects,
-  onCreated,
-}: {
-  open: boolean;
-  onClose: () => void;
-  projects: { id: string; name: string }[];
-  onCreated: () => void;
-}) {
-  const { user } = useAuth();
-  const [form, setForm] = useState<NewPurchaseForm>(EMPTY_FORM);
-  const [saving, setSaving] = useState(false);
-  const [dateOpen, setDateOpen] = useState(false);
-
-  useEffect(() => { if (open) setForm(EMPTY_FORM); }, [open]);
-
-  const set = (field: keyof NewPurchaseForm, value: unknown) =>
-    setForm((prev) => ({ ...prev, [field]: value }));
-
-  const handleSubmit = async () => {
-    // Validações de campos obrigatórios — mensagens específicas para cada caso
-    const missing: string[] = [];
-    if (!form.project_id) missing.push('Obra');
-    if (!form.item_name.trim()) missing.push('Nome do item');
-
-    if (missing.length > 1) {
-      toast.error('Campos obrigatórios não preenchidos', {
-        description: `Preencha: ${missing.join(', ')}.`,
-      });
-      return;
-    }
-    if (missing.length === 1) {
-      const map: Record<string, { title: string; desc: string }> = {
-        'Obra': {
-          title: 'Selecione uma obra',
-          desc: 'Escolha a obra para a qual esta compra será solicitada.',
-        },
-        'Nome do item': {
-          title: 'Informe o nome do item',
-          desc: 'Descreva brevemente o que está sendo solicitado (ex: Cimento CP-II 50kg).',
-        },
-      };
-      const m = map[missing[0]];
-      toast.error(m.title, { description: m.desc });
-      return;
-    }
-
-    if (!user?.id) {
-      toast.error('Sessão expirada', {
-        description: 'Não foi possível identificar seu usuário. Faça login novamente para continuar.',
-      });
-      return;
-    }
-
-    setSaving(true);
-    try {
-      // planned_purchase_date é opcional. required_by_date é NOT NULL no banco:
-      // se o usuário informou uma data válida, usamos ela; caso contrário, default = hoje + 7 dias.
-      const planned = form.planned_purchase_date;
-      const plannedValid = planned instanceof Date && isValid(planned);
-      const plannedDate = plannedValid ? format(planned, 'yyyy-MM-dd') : null;
-      const requiredDate = plannedValid
-        ? format(planned, 'yyyy-MM-dd')
-        : format(addDays(new Date(), 7), 'yyyy-MM-dd');
-
-      const payload: ProjectPurchaseInsert = {
-        project_id: form.project_id,
-        created_by: user.id,
-        item_name: form.item_name.trim(),
-        category: form.category.trim() || null,
-        supplier_name: form.supplier_name.trim() || null,
-        estimated_cost: form.estimated_cost ? Number(form.estimated_cost.replace(',', '.')) : null,
-        planned_purchase_date: plannedDate,
-        required_by_date: requiredDate,
-        quantity: form.quantity ? Number(form.quantity) : 1,
-        unit: form.unit.trim() || 'un',
-        description: form.description.trim() || null,
-        notes: form.notes.trim() || null,
-        status: 'pending',
-      };
-      const { error } = await supabase.from('project_purchases').insert(payload);
-      if (error) throw error;
-      toast.success('Solicitação criada com sucesso!', {
-        description: `${form.item_name.trim()} foi adicionada ao calendário de compras.`,
-      });
-      onCreated();
-      onClose();
-    } catch (e) {
-      console.error('[CalendarioCompras] insert error:', e);
-      const message = e instanceof Error ? e.message : 'Erro desconhecido';
-      toast.error('Não foi possível criar a solicitação', {
-        description: `${message}. Tente novamente em alguns instantes.`,
-      });
-    } finally {
-      setSaving(false);
-    }
-  };
-
-  return (
-    <Dialog open={open} onOpenChange={(o) => { if (!o) onClose(); }}>
-      <DialogContent className="max-w-xl max-h-[90vh] overflow-y-auto">
-        <DialogHeader>
-          <DialogTitle className="flex items-center gap-2"><Plus className="h-4 w-4" /> Nova Solicitação de Compra</DialogTitle>
-        </DialogHeader>
-
-        <div className="grid gap-4 py-2">
-          {/* Obra — obrigatório */}
-          <div className="grid gap-1.5">
-            <Label className="text-sm font-medium">Obra <span className="text-destructive">*</span></Label>
-            <Select value={form.project_id} onValueChange={(v) => set('project_id', v)}>
-              <SelectTrigger className={cn('h-9', !form.project_id && 'border-destructive/50')}>
-                <SelectValue placeholder="Selecionar obra…" />
-              </SelectTrigger>
-              <SelectContent className="z-[300]">
-                {projects.map((p) => (
-                  <SelectItem key={p.id} value={p.id}>{p.name}</SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
-          </div>
-
-          {/* Item — obrigatório */}
-          <div className="grid gap-1.5">
-            <Label className="text-sm font-medium">Item / Produto <span className="text-destructive">*</span></Label>
-            <Input placeholder="Ex: Cimento CP-II, Vergalhão 10mm…" value={form.item_name}
-              onChange={(e) => set('item_name', e.target.value)} className="h-9" />
-          </div>
-
-          {/* Categoria + Fornecedor */}
-          <div className="grid grid-cols-2 gap-3">
-            <div className="grid gap-1.5">
-              <Label className="text-sm">Categoria</Label>
-              <Input placeholder="Ex: Estrutura, Elétrico…" value={form.category}
-                onChange={(e) => set('category', e.target.value)} className="h-9" />
-            </div>
-            <div className="grid gap-1.5">
-              <Label className="text-sm">Fornecedor</Label>
-              <Input placeholder="Nome do fornecedor" value={form.supplier_name}
-                onChange={(e) => set('supplier_name', e.target.value)} className="h-9" />
-            </div>
-          </div>
-
-          {/* Qtde + Unidade + Custo estimado */}
-          <div className="grid grid-cols-3 gap-3">
-            <div className="grid gap-1.5">
-              <Label className="text-sm">Quantidade</Label>
-              <Input type="number" min="0" placeholder="0" value={form.quantity}
-                onChange={(e) => set('quantity', e.target.value)} className="h-9" />
-            </div>
-            <div className="grid gap-1.5">
-              <Label className="text-sm">Unidade</Label>
-              <Input placeholder="un, kg, m²…" value={form.unit}
-                onChange={(e) => set('unit', e.target.value)} className="h-9" />
-            </div>
-            <div className="grid gap-1.5">
-              <Label className="text-sm">Custo Estimado (R$)</Label>
-              <Input type="number" min="0" step="0.01" placeholder="0,00" value={form.estimated_cost}
-                onChange={(e) => set('estimated_cost', e.target.value)} className="h-9" />
-            </div>
-          </div>
-
-          {/* Data planejada */}
-          <div className="grid gap-1.5">
-            <Label className="text-sm">Data Planejada de Compra</Label>
-            <Popover open={dateOpen} onOpenChange={setDateOpen}>
-              <PopoverTrigger asChild>
-                <Button variant="outline" className={cn('w-full h-9 justify-start font-normal text-sm', !form.planned_purchase_date && 'text-muted-foreground')}>
-                  <CalendarIcon className="mr-2 h-4 w-4" />
-                  {form.planned_purchase_date ? format(form.planned_purchase_date, "dd 'de' MMMM 'de' yyyy", { locale: ptBR }) : 'Selecionar data…'}
-                </Button>
-              </PopoverTrigger>
-              <PopoverContent className="w-auto p-0 z-[300]" align="start">
-                <CalendarPicker mode="single" selected={form.planned_purchase_date}
-                  onSelect={(d) => { set('planned_purchase_date', d); setDateOpen(false); }}
-                  locale={ptBR} initialFocus className="p-3 pointer-events-auto" />
-              </PopoverContent>
-            </Popover>
-          </div>
-
-          {/* Descrição + Observações */}
-          <div className="grid gap-1.5">
-            <Label className="text-sm">Descrição técnica</Label>
-            <Textarea placeholder="Especificações, normas, referências técnicas…" value={form.description}
-              onChange={(e) => set('description', e.target.value)} rows={2} className="text-sm resize-none" />
-          </div>
-          <div className="grid gap-1.5">
-            <Label className="text-sm">Observações</Label>
-            <Textarea placeholder="Urgência, ponto de entrega, contato do fornecedor…" value={form.notes}
-              onChange={(e) => set('notes', e.target.value)} rows={2} className="text-sm resize-none" />
-          </div>
-        </div>
-
-        <DialogFooter className="gap-2 sm:gap-0">
-          <Button variant="outline" onClick={onClose} disabled={saving}>Cancelar</Button>
-          <Button onClick={handleSubmit} disabled={saving} className="gap-1.5">
-            {saving ? 'Salvando…' : <><Plus className="h-4 w-4" /> Criar Solicitação</>}
-          </Button>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
-  );
-}
 
 // ─── Main page ────────────────────────────────────────────────────────────────
 export default function CalendarioCompras() {
